refactor(chart): extract tooltip methods into ChartTooltipMixin
